srv/requirements.txt
Flask == 2.2.2
Werkzeug == 2.2.2
python-can == 4.6.1

Flask-WTF
Flask-SQLAlchemy
Flask-Login
WTForms
email-validator
gunicorn
numpy<2